refactor(cli): extract doctor command into commands/doctor.rs
commands/mod.rs mixed the read-command handlers with the entire doctor
health-check subsystem. Move the CheckResult/CheckStatus/DoctorReport types,
the check_* helpers, build_report/format_report_human, and cmd_doctor into
commands/doctor.rs, along with their tests (now in doctor's own test module).
mod.rs re-exports cmd_doctor.

Pure code movement, no API change (all cartog tests pass). mod.rs drops from
3022 to 2255 lines.
